+++ src/views/participants/SetParticipants.vue
@@ -65,9 +65,24 @@
             },
 
             deleteParticipant(id) {
-                this.api.delete(id).then(response => {
-                    this.$router.push('set-participants')
-                    this.getParticipants();
+                swal({
+                    title: "Are you sure?",
+                    text: "Once deleted, you will not be able to recover this record",
+                    icon: "warning",
+                    buttons: true,
+                    dangerMode: true,
+                })
+                .then((willDelete) => {
+                    if (willDelete) {
+                        this.api.delete(id).then(response => {
+                            swal("Participant has been deleted!", {
+                                icon: "success",
+                            });
+                            this.$router.push('set-participants')
+                            this.getParticipants();
+                        });
+                        
+                    } 
                 });
             }
         }
